packui script to support new sections approach
