+++ .prettierrc.yaml
@@ -3,12 +3,11 @@ bracketSpacing: true
 endOfLine: lf
 htmlWhitespaceSensitivity: css
 insertPragma: false
-printWidth: 120
+printWidth: 140
 proseWrap: preserve
 quoteProps: as-needed
 rangeStart: 0
 requirePragma: false
-semi: true
 singleQuote: true
 tabWidth: 2
 trailingComma: none
