+++ src/commands/veloce/loadmodel.ts
@@ -1,19 +1,12 @@
+import { readFileSync } from 'fs'
 import * as os from 'os'
 import * as path from 'path'
-import { readFileSync } from 'fs'
-import { default as axios } from 'axios'
 import { flags, SfdxCommand } from '@salesforce/command'
 import { Messages } from '@salesforce/core'
 import { AnyJson } from '@salesforce/ts-types'
-import {
-  LegacyUiDefinition,
-  UiDef,
-  UiDefinition,
-  UiElement,
-  UiMetadata
-} from '../../shared/types/ui.types'
-import { getDirectoryNames, readFileSafe } from '../../shared/utils/common.utils'
-import { extractElementMetadata, toBase64 } from '../../shared/utils/ui.utils'
+import { default as axios } from 'axios'
+import { UiDef } from '../../shared/types/ui.types'
+import { UiDefinitionsBuilder } from '../../shared/utils/ui.utils'
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname)
@@ -87,8 +80,7 @@ export default class Org extends SfdxCommand {
     this.ux.log('PML Successfully Loaded!')
 
     // load UI
-    const dir = `models/${name}`
-    const uiDefs: UiDef[] = [...this.packUiDefinitions(dir), ...this.packLegacyUiDefinitions(dir)]
+    const uiDefs: UiDef[] = new UiDefinitionsBuilder('models', name, {}, this.ux).pack()
     try {
       await axios.post(`${backendUrl}/services/dev-override/model/${name}/ui`, JSON.stringify(uiDefs), { headers })
     } catch ({ data }) {
@@ -99,118 +91,5 @@ export default class Org extends SfdxCommand {
 
     // Return an object to be displayed with --json
     return { model: pml, ui: JSON.stringify(uiDefs) }
-  }
-
-  private packUiDefinitions(dir: string): UiDefinition[] {
-    const folders = getDirectoryNames(dir)
-
-    const uiDefinitions: UiDefinition[] = folders.reduce((acc, folder) => {
-      const uiDir = `${dir}/${folder}`
-      const metadataString = readFileSafe(`${uiDir}/metadata.json`)
-
-      if (!metadataString) {
-        return acc
-      }
-
-      const metadata: UiMetadata = JSON.parse(metadataString)
-      const { children, ...rest } = metadata
-
-      const uiDefinition: UiDefinition = {
-        ...rest,
-        children: children.map(childName => this.packUiElement(`${uiDir}/${childName}`))
-      }
-
-      return [...acc, uiDefinition]
-    }, [] as UiDefinition[])
-
-    return uiDefinitions
-  }
-
-  private packUiElement(dir: string): UiElement {
-    const script = readFileSafe(`${dir}/script.ts`)
-    const metadata = extractElementMetadata(script)
-
-    const element: UiElement = {
-      script: toBase64(script),
-      children: metadata.children.map(childName => this.packUiElement(`${dir}/${childName}`))
-    }
-
-    const styles = readFileSafe(`${dir}/styles.css`)
-    if (styles) {
-      element.styles = toBase64(styles)
-    }
-
-    const template = readFileSafe(`${dir}/template.html`)
-    if (template) {
-      element.template = toBase64(template)
-    }
-
-    return element
-  }
-
-  private packLegacyUiDefinitions(dir: string): LegacyUiDefinition[] {
-    const metadataString = readFileSync(`${dir}/metadata.json`, 'utf-8')
-
-    if (!metadataString) {
-      return []
-    }
-
-    const legacyDefinitions: LegacyUiDefinition[] = JSON.parse(metadataString)
-
-    for (const ui of legacyDefinitions) {
-      for (const section of ui.sections) {
-        if (section.templateUrl) {
-          const p = `models/${section.templateUrl.trim()}`
-          this.assertPath(p)
-          const b = readFileSync(p)
-          const base64 = b.toString('base64')
-          section.template = base64
-          delete section.templateUrl
-        }
-        if (section.scriptUrl) {
-          const p = `models/${section.scriptUrl.trim()}`
-          this.assertPath(p)
-          const b = readFileSync(p)
-          const base64 = b.toString('base64')
-          section.script = base64
-          delete section.scriptUrl
-        }
-        if (section.stylesUrl) {
-          const p = `models/${section.stylesUrl.trim()}`
-          this.assertPath(p)
-          const b = readFileSync(p)
-          const base64 = b.toString('base64')
-          section.styles = base64
-          delete section.stylesUrl
-        }
-        if (section.propertiesUrl) {
-          const p = `models/${section.propertiesUrl.trim()}`
-          this.assertPath(p)
-          section.properties = this.parseJsonFile(p)
-          delete section.propertiesUrl
-        }
-      }
-    }
-
-    return legacyDefinitions
-  }
-
-  private assertPath(p: string): void {
-    for (const part of p.split('/')) {
-      if (part.startsWith(' ') || part.endsWith(' ') || part.startsWith('\t') || part.endsWith('\t')) {
-        this.ux.log(`Path has leading trailing/leading spaces, please remove and rename folder: ${p}`)
-        process.exit(255)
-      }
-    }
-  }
-
-  private parseJsonFile(p: string): any {
-    try {
-      const b = readFileSync(p).toString()
-      return JSON.parse(b)
-    } catch (e) {
-      this.ux.log('Failed to read/parse JSON file ', e)
-      process.exit(255)
-    }
   }
 }
